pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
backend-path = ["."]

[tool.hatch.version]
path = "yolo_ncnn/__init__.py"

[tool.hatch.build]
exclude = [
    "results/",
    "images/",
    "models/",
    ".pre-commit-config.yaml",
    ".vscode/",
    ".github/",
    "!yolo_ncnn",
]

[project]
name = "yolo_ncnn"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ncnn>=1.0.20250503",
    "pillow>=11.3.0",
    "pyyaml>=6.0.2",
]


[tool.ruff]
line-length = 120
lint.select = [
    "B",   # flake8-bugbear
    "D",   # pydocstyle
    "E",   # pycodestyle 错误
    "F",   # pyflakes
    "I",   # isort
    "SIM", # flake8-simplify
    "W",   # pycodestyle 警告
    "UP",  # pyupgrade
    "FA",  # flake8-future-annotationscl
    "ISC", # flake8-implicit-str-concat
    "ICN", # flake8-import-conventions
    "PTH", # flake8-use-pathlib
    "NPY", # numpy

]
lint.ignore = [
    "B008", # 在函数调用中使用函数调用作为默认参数
    "D100", # 缺少模块级 docstring
    "D101", # 缺少类级 docstring
    "D105", # 缺少魔术方法 docstring
    "D203", # 需要在类 docstring 之前至少有一个空行
    "D213", # 多行docstring的摘要应该从第二行开始
    "D401", # 第一行的docstring应该是命令式的
    "D406", # 节名应该以换行符结束
    "D407", # 节名后未出现虚线
    "E501", # 超长的行
    "F403", # 使用 'from module import *' 导入，无法检测未定义的名称
    "F405", # 名称可能未定义，或者是从星号导入的
    "F541", # f-string中没有占位符
]
lint.pydocstyle.convention = "google"

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["D104", "F401"]
"*.pyi" = ["D301", "D415", "D417", "D418", "E742"]
"tests/*.py" = ["D"]

[[tool.uv.index]]
url = "https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple/"
default = true

[dependency-groups]
dev = ["pytest>=8.4.1"]
